texture update & code
added 3d model
updated light source
created shop system

+++ Code/Tower Defence Protocall/Assets/scripts/monsterScripts/Monster.cs
@@ -11,7 +11,7 @@ public class Monster : MonoBehaviour
     public float maxHealth;
     public float remainingHealth;
     public float damageDealt;
-    public float moneyDrop;
+    public int moneyDrop;
     public float speed;
     public float difficulty;
